@@ -12,9 +12,9 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@types/node": "^13.5.0",
+    "@types/node": "^13.13.2",
     "ts-clean": "^1.0.3",
-    "typescript": "^3.7.5"
+    "typescript": "^3.8.3"
   },
   "dependencies": {
     "@aws-cdk/aws-apigateway": "^1.27.0",
