@@ -11,18 +11,18 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "aws-cdk": "^1.18.0",
-    "@aws-cdk/assert": "^1.18.0",
-    "@types/jest": "^24.0.22",
+    "@aws-cdk/assert": "^1.35.0",
+    "@types/jest": "^24.9.1",
     "@types/node": "10.17.5",
+    "aws-cdk": "^1.35.0",
     "jest": "^24.9.0",
-    "ts-jest": "^24.1.0",
-    "ts-node": "^8.1.0",
-    "typescript": "~3.7.2"
+    "ts-jest": "^24.3.0",
+    "ts-node": "^8.9.0",
+    "typescript": "^3.8.3"
   },
   "dependencies": {
-    "@aws-cdk/aws-dynamodb": "^1.18.0",
-    "@aws-cdk/core": "^1.18.0",
-    "cdk-dynamo-table-viewer": "^3.0.6"
+    "@aws-cdk/aws-dynamodb": "^1.35.0",
+    "@aws-cdk/core": "^1.35.0",
+    "cdk-dynamo-table-viewer": "^3.1.2"
   }
 }
